Unify token refresh logic with proactive interceptor, timeout, and retryWithBackoff
- Integrated tokenRefreshManager into requestHandler for proactive refresh checks
- Added 15-second timeout to all token refresh calls with AbortController
- Refactored tokenRefreshManager to use retryWithBackoff helper (eliminated duplication)
- Consolidated refreshPromise deduplication to single source in tokenRefreshManager
- Simplified authService.refreshAccessToken to thin wrapper calling performTokenRefresh
