+++ frontend/services/authService.ts
@@ -37,7 +37,7 @@ import {
   SecureAuthData,
 } from "@/utils/authTokenStore";
 import { clearAllExchangedTokens } from "@/utils/exchangedTokenStore";
-import { showLogoutConfirmation, showNetworkError, showRefreshRetryDialog } from "@/utils/authAlerts";
+import { performTokenRefresh, AuthData } from "@/utils/tokenRefreshManager";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import axios from "axios";
 import dayjs from "dayjs";
@@ -46,27 +46,16 @@ import { Alert } from "react-native";
 import { logout as appLogout, AuthorizeResult } from "react-native-app-auth";
 
 const GRANT_TYPE_AUTHORIZATION_CODE = "authorization_code";
-const GRANT_TYPE_REFRESH_TOKEN = "refresh_token";
 const GRANT_TYPE_TOKEN_EXCHANGE =
   "urn:ietf:params:oauth:grant-type:token-exchange";
 const SUBJECT_TOKEN_TYPE = "urn:ietf:params:oauth:token-type:jwt";
 const REQUESTED_TOKEN_TYPE = "urn:ietf:params:oauth:token-type:access_token";
 const MILLISECONDS_IN_A_SECOND = 1000;
 const SCOPE = "openid email groups profile";
-let refreshPromise: Promise<AuthData | null> | null = null;
 
 export interface DecodedIdToken {
   email?: string;
   userid?: string;
-}
-
-export interface AuthData {
-  accessToken: string;
-  refreshToken: string;
-  idToken: string;
-  email?: string;
-  userId?: string;
-  expiresAt: number;
 }
 
 export const getAccessToken = async (
@@ -126,117 +115,7 @@ export const getAccessToken = async (
 export const refreshAccessToken = async (
   onLogout: () => Promise<void>
 ): Promise<AuthData | null> => {
-  if (refreshPromise) {
-    return refreshPromise; // If a refresh is already in progress, return the same promise
-  }
-
-  refreshPromise = (async () => {
-    try {
-      const storedData = await loadAuthDataFromSecureStore();
-      if (!storedData) {
-        refreshPromise = null;
-        return null;
-      }
-
-      const authData: AuthData = storedData;
-      if (!authData.refreshToken) {
-        refreshPromise = null;
-        return null;
-      }
-
-      if (!TOKEN_URL) {
-        console.error(
-          "TOKEN_URL is not defined. Check your environment variables."
-        );
-        refreshPromise = null;
-        return null;
-      }
-
-      const requestBody = createAuthRequestBody({
-        grantType: GRANT_TYPE_REFRESH_TOKEN,
-        clientId: CLIENT_ID,
-        refreshToken: authData.refreshToken,
-      });
-
-      const response = await fetch(TOKEN_URL, {
-        method: "POST",
-        headers: { "Content-Type": "application/x-www-form-urlencoded" },
-        body: requestBody,
-      });
-
-      if (!response.ok) {
-        console.error(
-          `Token refresh failed: ${response.status} ${response.statusText}`
-        );
-        if (response.status === 400) {
-          await showLogoutConfirmation(
-            "Session Expired",
-            "Your session has expired. Would you like to sign in again?",
-            onLogout
-          );
-        }
-
-        refreshPromise = null;
-        return null;
-      }
-
-      const data = await response.json();
-
-      if (data.access_token && data.id_token) {
-        const decodedIdToken = jwtDecode<DecodedIdToken>(data.id_token);
-        const exp = jwtDecode<{ exp: number }>(data.access_token).exp || 0;
-        const { email, userid: userId } = decodedIdToken;
-
-        const updatedAuthData: AuthData = {
-          accessToken: data.access_token,
-          refreshToken: data.refresh_token || authData.refreshToken,
-          idToken: data.id_token,
-          email,
-          userId,
-          expiresAt: exp * MILLISECONDS_IN_A_SECOND,
-        };
-
-        await saveAuthDataToSecureStore(updatedAuthData as SecureAuthData);
-
-        refreshPromise = null;
-        return updatedAuthData;
-      }
-
-      refreshPromise = null;
-      return null;
-    } catch (err: unknown) {
-      if (err instanceof Error) {
-        console.error("Token refresh error:", err.message);
-        if (err.message.includes("Network request failed") || err.message.includes("timeout")) {
-          await showNetworkError("Check your connection and try again.");
-        } else {
-          await showRefreshRetryDialog(
-            "Error",
-            "An error occurred while refreshing your session. Would you like to try again or sign in?",
-            async () => {
-              await refreshAccessToken(onLogout);
-            },
-            onLogout
-          );
-        }
-      } else {
-        console.error("An unexpected error occurred during token refresh.");
-        await showRefreshRetryDialog(
-          "Error",
-          "An unexpected error occurred. Would you like to try again or sign in?",
-          async () => {
-            await refreshAccessToken(onLogout);
-          },
-          onLogout
-        );
-      }
-
-      refreshPromise = null;
-      return null;
-    }
-  })();
-
-  return refreshPromise;
+  return performTokenRefresh(onLogout);
 };
 
 export const logout = async () => {
