+++ frontend/utils/requestHandler.ts
@@ -13,31 +13,22 @@
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
-import { refreshAccessToken } from "@/services/authService";
+import { checkAndRefreshTokenIfNeeded } from "@/utils/tokenRefreshManager";
 import { loadAuthDataFromSecureStore } from "@/utils/authTokenStore";
+import { refreshAccessToken } from "@/services/authService";
 import axios, { AxiosRequestConfig } from "axios";
-import dayjs from "dayjs";
-import { jwtDecode } from "jwt-decode";
 
 // General API request handler
 export const apiRequest = async (
   config: AxiosRequestConfig,
   onLogout: () => Promise<void>
 ) => {
-  let accessToken = await getAccessToken(); // Get stored access token
+  let accessToken = await getAccessToken();
 
-  // If no access token, return early
   if (!accessToken) return;
 
-  // Check if token is expired before making request
-  if (isAccessTokenExpired(accessToken)) {
-    const newAuthData = await refreshAccessToken(onLogout);
-
-    if (!newAuthData?.accessToken) {
-      return; // Logout is triggered inside refreshAccessToken
-    }
-    accessToken = newAuthData.accessToken;
-  }
+  await checkAndRefreshTokenIfNeeded();
+  accessToken = await getAccessToken();
 
   // Set Authorization Header
   config.headers = {
@@ -74,18 +65,7 @@ export const apiRequest = async (
   }
 };
 
-// Helper function to check if the token is expired
-const isAccessTokenExpired = (accessToken: string): boolean => {
-  try {
-    const decoded = jwtDecode<{ exp: number }>(accessToken);
-    return dayjs.unix(decoded.exp).isBefore(dayjs());
-  } catch {
-    return true; // Assume expired if decoding fails
-  }
-};
-
-// Helper function to check if error is retryable
-const isRetryableError = (error: unknown): boolean => {
+export const isRetryableError = (error: unknown): boolean => {
   if (error instanceof Error) {
     return (
       error.message.includes("Network request failed") ||
@@ -101,7 +81,8 @@ export async function retryWithBackoff<T>(
   fn: () => Promise<T>,
   maxRetries = 3,
   baseDelay = 100,
-  multiplier = 2
+  multiplier = 2,
+  jitterMs = 0
 ): Promise<T> {
   for (let attempt = 1; attempt <= maxRetries; attempt++) {
     try {
@@ -112,7 +93,8 @@ export async function retryWithBackoff<T>(
       }
 
       const delay = baseDelay * Math.pow(multiplier, attempt - 1);
-      await new Promise(resolve => setTimeout(resolve, delay));
+      const jitter = jitterMs > 0 ? Math.random() * jitterMs : 0;
+      await new Promise(resolve => setTimeout(resolve, delay + jitter));
     }
   }
   throw new Error('Max retries exceeded');
